// src/app/models/repoData.ts
/* 
created_at
description
forks_count
language
name
pushed_at
stargazers_count
topics
updated_at
watchers_count
html_url
size (kilobytes)
*/

export type repoData = {
  name: string;
  description: string;
  html_url: string;
  created_at: string;
  updated_at: string;
  pushed_at: string;
  forks_count: number;
  stargazers_count: number;
  watchers_count: number;
  size: number;
  language: string;
};